Ansible 12 compat
src may be _OmitType if the first candidate is not found. first_found
needs a list of strings. _OmitType also does not have a length.

Instead use NoneType and remove it from the list of candidates. The
result ist either a src String that was found, or an empty list in which
case the test ensures the module is skipped.