Ansible <12 and >=12 compat
Ansible < 12:

`default(None)` is mapped to an empty string which causes `first_found`
to find a directory. Which cannot be used to template and an error is
raised.

By applying `select` without a test argument we remove empty strings for
the list of candidates.

Ansible >= 12:

default(None) is mapped to null which causes first_found to ignore the
list's first entry.

`select` without argument removes the null element before `first_found`
receives it as input (and would ignore it anyway).